spdm: add ReadHalf and ReadHalf and into_split()
We have found that the SpdmRequester and SpdmRequester are hard to use. Specifically speaking, it is hard to send message to peer while receiving message from peer at the same time and can easily lead to deadlocks.

In this commit, add a `into_split()` function (inspired by [tokio](https://docs.rs/tokio/latest/tokio/net/struct.TcpStream.html#method.into_split)), to split SpdmRequester or SpdmRequester to a `(ReadHalf,ReadHalf)`.

Note that this commit leads to a tight coupling with TcpStream to some extent (e.g. call on TcpStream::try_clone()). This is caused by the design flaws in the SpdmDeviceIo in spdm-rs. The future work requires a refactoring of SpdmDeviceIo.

Signed-off-by: Kun Lai <me@imlk.top>