rats-rs: split transport layer to rats-transport for ignoring spdm git submodule
In the previous version we have to init the submodule 'spdm-rs' even
though we are not using the transport layer (e.g. c-api).

In this commit we move the rats-rs/transport module to a single crate
rats-transport. And we put rats-transport and examples/spdm to a
standalone cargo workspace since the [patch] node can only be put in
workspace Cargo.toml.

Signed-off-by: Kun Lai <laikun@linux.alibaba.com>